Changed how arrows appear (WIP), added penalty for missing arrows and mispressing keys.
The arrows appear right next to the player and if there are multiple arrows of the same direction, they spawn at an offset. They disappear after a certain amount of time or if the correct input key is being pressed.

TO CHANGE: The arrow furthest to player should be removed first.

BUG: When correct input key is being pressed, every arrow of the same direction will get deleted and added to score counter.